Improve support for media keys: next, previous, play
MEDIA_PREVIOUS and MEDIA_NEXT cycle through the list of favourites and
start playing. If the current or previously played station is not in the
favourites list, the keys will start playing the last or the first entry
from the favourites list, respectively.

Media keys now also work when the app is in the background. This is also
the case for keys that trigger onPlay. The latter will play the previous
station from the history list as fallback.