Fix long text being truncated in BottomNavigationView
It was a regression introduced when I updated the libraries in 8ccb828.

Now it needs to be explicitly specified that selected item should push
other aside in BottomNavigationView.

In some cases it still may not be enough to fit the text but it doesn't
happen at the moment so nothing more is needed to be done. Otherwise
it will be necessary to make the TextViews to auto resize the text.

Closes: #607
Fixes: 8ccb828