Introduce Espresso test framework and several general tests
It's time to add some tests.

Our code was not made with unit tests in mind and most of the errors are
in the interaction between components, so testing UI looks like a way to go.

Each test will be launched with a clean state. The drawback is that each test
takes considerable amount of time - several seconds.

Http requests are mocked with predefined content, if there is no content
- the 404 code is returned.