Show notification if service is started by external forces
It is an error for service started via Context.startForegroundService
not to create a notification since Android O. It must call startForeground
within 5 seconds of being started.
Thus if we can show a notification - we always show it.

However we don't show notification if service is started from our app -
since we always start it when app is launched.

Closes: #634
Closes: #619