Fix bottom sheet positioning on Android Q
Android Q introduced gesture navigation which in combination with
edge-to-edge app mode may cause gestures conflict with bottom sheet.

To fix this in new release of material components
bottom sheet automatically adjust position based on system defined
insets. However it doesn't know where bottom sheet actually is
and if the app supports edge-to-edge mode.

Since we don't support edge-to-edge mode - we could just disable
this new feature of bottom sheet.

Closes: #798