Add setting for icon display in media browser
According to android-media-controller auto test suite¹, icon bitmaps for
the media browser must be null and icon uris are restricted to content
uris and android resources. So there is no  supported way of using icons
here.
However, icons in the browser are extremely helpful, especially while
driving and network use is also no issue in this case – and the icons
work. Thus, introduce a settings and let the user decide.

¹https://github.com/googlesamples/android-media-controller/blob/0bbad7b22ca2c087629cce0e5b0eb61f07f3fde1/mediacontroller/src/main/java/com/example/android/mediacontroller/MediaAppTestDetails.kt#L1226