Android TV: show starred stations as a home screen channel
TvChannelManager is an Observer publishing starred programmes
as a home screen channel.

The channel is created as non-browseable by default since when
a channel doesn’t have any programmes, it shows a button launching
the application itself — probably not what a user wants if they never
used the app and never starred any stations.

As a first station is starred, TvChannelManager requests the system
to make the channel browseable. This works for the first channel
the application creates *only*, which is why the channel is never
removed — if we recreate a channel, it won’t be shown on the home
screen and the user will have to make it visible manually.

The APIs this code uses are available since the API level 26 only,
and the library wrapping them needs minimal SDK version 21; however,
since we want to support API level 16, this needs to be overridden
in the manifest.

Signed-off-by: Andrej Shadura <andrew.shadura@collabora.co.uk>